calcul du spacing en fonction de la map

+++ srcs/debug.c
@@ -6,7 +6,7 @@
 /*   By: nbellila <nbellila@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/29 20:15:11 by nbellila          #+#    #+#             */
-/*   Updated: 2024/07/02 03:51:01 by nbellila         ###   ########.fr       */
+/*   Updated: 2024/07/03 13:16:55 by nbellila         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	show_map(t_map *map)
 {
-	size_t	col;
-	size_t	row;
+	int	col;
+	int	row;
 
 	row = 0;
 	while (row < map->height)
